requirements.txt
boto3>=1.34.86
python-dotenv>=1.0.0
